trinetra_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
pydantic==2.7.1
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
